+++ src/ulysses-omnichain/interfaces/IERC20hTokenRoot.sol
@@ -15,7 +15,7 @@ interface IERC20hTokenRoot {
     //////////////////////////////////////////////////////////////*/
 
     /// @notice View Function returns Local Network Identifier.
-    function localChainId() external view returns (uint256);
+    function localChainId() external view returns (uint16);
 
     /// @notice View Function returns Root Port Address.
     function rootPortAddress() external view returns (address);
